python/playground_requirements.txt
tika==2.6.0
playwright
python-dotenv
beautifulsoup4
tqdm
scraperapi-sdk
jsonschema
tqdm
aiomysql